Add new workerbee2 behaviour
Now supports a number of new fields for computing statistics
about the job completion behaviour. Rather than booleans
for job completion and job claiming we now store timestamps.
We also prefer the OLDEST job rather than a random one.

We use auto-increment ids for the job ids and thus support easily
adding new jobs. A new unique_id field is maintained for the
previous use case of those who want to use it to store data
rather than the input_data field.

The previously optional data field becomes the input_data json
field which is used to store job data. This is matched
with the new output_data field (optional) for storing the results
of a job.

We also keep track of how long a successful job took.

Finally, we add two new binary exponential backoff strategies -
one for failures and one for 'busywaiting'. If busywait is
enabled, then when all jobs are completed the system will
sleep in increasing intervals whilst waiting for a new job
to appear. If a job fails, the system will sleep increasing
amounts until it recieves a non-failing job. There is also
a maximum number of retries and jobs larger than this value
will not be considered as available jobs.